:sparkles: #10 収入履歴のスキーマを別ファイルへ移動
